requirements.txt
discord.py
asyncio
aiosqlite
requests
Pillow
psutil
